requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
backoff==2.2.1
Bottleneck==1.4.2
brotlicffi==1.2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
einops==0.8.2
emcee==3.1.6
eval_type_backport==0.3.1
exceptiongroup==1.3.1
filelock==3.24.3
fonttools==4.61.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numexpr==2.14.1
numpy==2.2.5
nvidia-ml-py==13.590.48
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.2
posthog==6.9.3
progress==1.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
random2==1.0.2
requests==2.32.5
rich==14.3.3
ruff==0.15.2
scikit-learn==1.7.2
scipy==1.15.3
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tabpfn==6.4.1
tabpfn-common-utils==0.2.16
threadpoolctl==3.5.0
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wheel==0.46.3
